buildFromCache.test file removed after its contents were placed in QuellCache.test
